Refactor: Extract plan subscription logic
Separate plan subscription logic from `AvailablePlansList.tsx` into a dedicated hook to improve component focus and readability.
